// lib/type-defs.graphqls.d.ts
export declare type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
};
export declare type Query = {
    __typename?: 'Query';
    viewer: User;
    viewerRemote: Array<Maybe<UserRemote>>;
};
export declare type User = {
    __typename?: 'User';
    id: Scalars['ID'];
    name: Scalars['String'];
    status: Scalars['String'];
};
export declare type UserRemote = {
    __typename?: 'UserRemote';
    id: Scalars['ID'];
    name: Scalars['String'];
    status: Scalars['String'];
};